v0.0.1: Cut/paste, Windows fixes, testing support
Ctrl-Y will now paste text cut with Ctrl-K and Ctrl-U.

Repl.run() is now synchronous. Use Repl.runAsync() for an asynchronous
version.

When the terminal doesn't support ANSI escaped input, it should fallback
to a line-by-line reader that will work on Windows.

When no terminal is present, Repl.run() prints the input along with
the output, allowing for easier testing.